fix: resolve critical issues in workersbd repository
Move misplaced page files and fix dynamic route naming conflict
Create missing components and fix import paths for dynamic pages
Update Next.js routing and fix ESLint errors
