+++ components/SEOHead.jsx
@@ -1,0 +1,46 @@
+import Head from 'next/head';
+
+export default function SEOHead({ 
+  title = 'WorkersBD - Find Jobs & Skilled Workers in Bangladesh',
+  description = 'Bangladesh\'s leading job portal. Find quality jobs and skilled workers in Dhaka, Chittagong, Sylhet, and across all districts.',
+  canonical = 'https://workersbd.com',
+  ogImage = 'https://workersbd.com/og-image.jpg',
+  locale = 'en',
+  schemas = []
+}) {
+  return (
+    <Head>
+      <title>{title}</title>
+      <meta name="description" content={description} />
+      <link rel="canonical" href={canonical} />
+      
+      {/* Open Graph */}
+      <meta property="og:title" content={title} />
+      <meta property="og:description" content={description} />
+      <meta property="og:image" content={ogImage} />
+      <meta property="og:url" content={canonical} />
+      <meta property="og:type" content="website" />
+      <meta property="og:locale" content={locale === 'bn' ? 'bn_BD' : 'en_US'} />
+
+      {/* Twitter Card */}
+      <meta name="twitter:card" content="summary_large_image" />
+      <meta name="twitter:title" content={title} />
+      <meta name="twitter:description" content={description} />
+      <meta name="twitter:image" content={ogImage} />
+
+      {/* Additional Meta Tags */}
+      <meta name="theme-color" content="#0066cc" />
+      <meta name="apple-mobile-web-app-capable" content="yes" />
+      <meta name="apple-mobile-web-app-status-bar-style" content="black-translucent" />
+
+      {/* Structured Data */}
+      {schemas.map((schema, index) => (
+        <script
+          key={index}
+          type="application/ld+json"
+          dangerouslySetInnerHTML={{ __html: JSON.stringify(schema) }}
+        />
+      ))}
+    </Head>
+  );
+}
