+++ lib/analytics.js
@@ -148,7 +148,7 @@ export const trackConversion = (type, value) => {
   fbEvent('Purchase', { value: value, currency: 'BDT' });
 };
 
-export default {
+const analyticsModule = {
   pageview,
   event,
   trackJobView,
@@ -164,3 +164,5 @@ export default {
   getVariant,
   trackConversion,
 };
+
+export default analyticsModule;
