+++ lib/seo.js
@@ -283,7 +283,7 @@ export function generateMetaTags({
   };
 }
 
-export default {
+const seoModule = {
   siteConfig,
   generateJobPostingSchema,
   generatePersonSchema,
@@ -295,3 +295,5 @@ export default {
   bangladeshDistricts,
   jobCategories
 };
+
+export default seoModule;
