+++ public/robots.txt
@@ -1,11 +1,25 @@
-# robots.txt for WorkersBD
+# WorkersBD Robots Configuration
+# Allow all crawlers except bad actors
+
 User-agent: *
 Allow: /
-
-# Block internal API and admin routes
-Disallow: /api/
 Disallow: /admin/
+Disallow: /api/
 Disallow: /_next/
+Disallow: /.git/
 
-# Sitemap
+# Sitemap location
 Sitemap: https://workersbd.com/sitemap.xml
+
+# Crawl delay for performance
+Crawl-delay: 1
+
+# Google specific
+User-agent: Googlebot
+Allow: /
+Crawl-delay: 0
+
+# Bing specific
+User-agent: Bingbot
+Allow: /
+Crawl-delay: 1
